Add asynchronous mode to methods.
Every method which acts on EC2 can now accept an optional parameter
'CallBack' which is a function to be called when the request is complete.

This callback method will be provided in the first argument with the same
data the function would have provided in it's return value had it been called
without the CallBack argument.

The return value when a CallBack argument is specified will instead be an
AnyEvent::Condvar which can be waited upon using $condvar->recv() to know
when the callback has been completed.